Remove video output from glut loop.
This is intended to decouple the simulation from the video output as both parts are compicated and make debugging much harder. This change can be followed by making the main project output to a state-based save format. Then another project can be created that plays back the save and possiby outputs it to a video file.