Refactor code to decouple render from project specific requirements
render.h & render.cpp should be usable across projects to eliminate code duplication. To do this, change render_init to accept the callbacks rather than provide them internally. This change also facilitates moving away from a global ParticleSystem. Instead, it is local to main and a global pointer to it is used.