Refactor some OMP loops to work around an MSVC compiler bug
The code is equivalent. In theory it could incur an additional runtime overhead because it breaks the big parallel region into smaller ones, but any sane compiler should optimize it anyway and keep the threads alive. It can be reverted at a later time if the compiler bug is fixed.