Replace visual studio solution with cmake files
This is the first in a possibly long line of WIP commits where the
project won't build. They are used as a way to backup progress and
should stay out of the master branch. The end goal is a project
that builds on both linux and windows. Other OSs would be good but
I probably won't make the effort to test that.